Movement now works BUT (read extended desc...)
Movement works using the arrow keys and 'a' and 'd' but when I try to
move (at least on my computer) ticks are running like, a billion times a
second and the image flies off the screen.

I confirmed this by adding a "sysout" function call to print "moved" to
the console whenever I hit 'a' and it only ran about a thousand times
per second, significantly slowing down the movement of the tiles and
allowing me to see what was happening. I then hit 'd' and everything
flew off the screen again.

I'll add a game loop tomorrow. :D Also, I'm making this an issue becuz I
want this project to "have issues."